Prepare for multiple frameworks support.
