+++ src/manager/ColorIcon.tsx
