+++ src/parameters.ts
@@ -1,0 +1,8 @@
+import { PARAM_KEY } from './constants';
+
+const parameters = {
+  name: 'withThemes',
+  parameterName: PARAM_KEY,
+} as const;
+
+export default parameters;
