+++ src/register.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { addons, types } from '@storybook/addons';
 
 import { ADDON_ID } from './constants';
-import { ThemeSelector } from './containers';
+import { ThemeSelector } from './manager';
 
 addons.register(ADDON_ID, api => {
   addons.add(ADDON_ID, {
